# openapi.json
{"openapi":"3.1.0","info":{"title":"Booskiff","description":"Booskiff drive API: per-owner file and folder storage with streaming uploads, publishing, and download URLs, built on the billing policy foundation (plans, rules, quotas) that governs storage limits.","license":{"name":"MIT","identifier":"MIT"},"version":"0.1.0"},"paths":{"/healthz":{"get":{"tags":["health"],"operationId":"healthz","responses":{"200":{"description":""}}}},"/public/{key}":{"get":{"tags":["public"],"operationId":"get_public_file","parameters":[{"name":"key","in":"path","description":"Public key of a published file","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The file content, served with immutable caching"},"404":{"description":"Unknown or unpublished key"}}}},"/readyz":{"get":{"tags":["health"],"operationId":"ready","responses":{"200":{"description":""},"503":{"description":""}}}},"/v1/admin/billing/rules":{"get":{"tags":["admin"],"operationId":"list_billing_rules","responses":{"200":{"description":"All billing rules, global layer first","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BillingRuleList"}}}},"401":{"description":"Missing, unknown, or revoked admin token"}},"security":[{"admin_token":[]}]},"post":{"tags":["admin"],"operationId":"create_billing_rule","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRuleRequest"}}},"required":true},"responses":{"200":{"description":"Rule inserted or updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BillingRuleItem"}}}},"400":{"description":"Partial owner scope, unknown key, or invalid value"},"401":{"description":"Missing, unknown, or revoked admin token"}},"security":[{"admin_token":[]}]}},"/v1/admin/billing/rules/{id}":{"delete":{"tags":["admin"],"operationId":"delete_billing_rule","parameters":[{"name":"id","in":"path","description":"Billing rule id","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"Rule deleted"},"401":{"description":"Missing, unknown, or revoked admin token"},"404":{"description":"Unknown rule id"}},"security":[{"admin_token":[]}]}},"/v1/admin/owners/{owner_type}/{owner_id}/plan":{"get":{"tags":["admin"],"operationId":"get_plan","parameters":[{"name":"owner_type","in":"path","description":"Owner type, e.g. account","required":true,"schema":{"type":"string"}},{"name":"owner_id","in":"path","description":"Owner id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Assigned plan; default when unassigned","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlanResponse"}}}},"401":{"description":"Missing, unknown, or revoked admin token"}},"security":[{"admin_token":[]}]},"put":{"tags":["admin"],"operationId":"set_plan","parameters":[{"name":"owner_type","in":"path","description":"Owner type, e.g. account","required":true,"schema":{"type":"string"}},{"name":"owner_id","in":"path","description":"Owner id","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetPlanRequest"}}},"required":true},"responses":{"204":{"description":"Plan assigned, replacing any previous assignment"},"400":{"description":"Unknown plan name"},"401":{"description":"Missing, unknown, or revoked admin token"}},"security":[{"admin_token":[]}]},"delete":{"tags":["admin"],"operationId":"delete_plan","parameters":[{"name":"owner_type","in":"path","description":"Owner type, e.g. account","required":true,"schema":{"type":"string"}},{"name":"owner_id","in":"path","description":"Owner id","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Assignment removed"},"401":{"description":"Missing, unknown, or revoked admin token"},"404":{"description":"No assignment for this owner"}},"security":[{"admin_token":[]}]}},"/v1/admin/owners/{owner_type}/{owner_id}/usage":{"get":{"tags":["admin"],"operationId":"get_usage","parameters":[{"name":"owner_type","in":"path","description":"Owner type, e.g. account","required":true,"schema":{"type":"string"}},{"name":"owner_id","in":"path","description":"Owner id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Metered received bytes; 0 when never uploaded","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsageResponse"}}}},"401":{"description":"Missing, unknown, or revoked admin token"}},"security":[{"admin_token":[]}]}},"/v1/admin/tokens":{"get":{"tags":["admin"],"operationId":"list_tokens","responses":{"200":{"description":"All admin tokens without secret material","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminTokenList"}}}},"401":{"description":"Missing, unknown, or revoked admin token"}},"security":[{"admin_token":[]}]},"post":{"tags":["admin"],"operationId":"create_token","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTokenRequest"}}},"required":true},"responses":{"201":{"description":"Token created; the raw token is returned exactly once","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatedAdminToken"}}}},"401":{"description":"Missing, unknown, or revoked admin token"}},"security":[{"admin_token":[]}]}},"/v1/admin/tokens/{id}":{"delete":{"tags":["admin"],"operationId":"revoke_token","parameters":[{"name":"id","in":"path","description":"Admin token id","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"Revoked; the token authenticates no longer"},"401":{"description":"Missing, unknown, or revoked admin token"},"404":{"description":"Unknown token id"}},"security":[{"admin_token":[]}]}},"/v1/billing/status":{"get":{"tags":["billing"],"operationId":"get_billing_status","responses":{"200":{"description":"Usage and effective limits for the authenticated owner","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BillingStatusResponse"}}}}},"security":[{"bearer_auth":[]}]}},"/v1/files":{"get":{"tags":["files"],"operationId":"list_files","responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileListResponse"}}}}},"security":[{"bearer_auth":[]}]},"post":{"tags":["files"],"operationId":"upload_file","responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileResponse"}}}}},"security":[{"bearer_auth":[]}]}},"/v1/files/{id}":{"get":{"tags":["files"],"operationId":"get_file","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileResponse"}}}},"404":{"description":""}},"security":[{"bearer_auth":[]}]},"delete":{"tags":["files"],"operationId":"delete_file","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":""},"404":{"description":""}},"security":[{"bearer_auth":[]}]}},"/v1/files/{id}/download-url":{"get":{"tags":["files"],"operationId":"download_url","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UrlResponse"}}}},"404":{"description":""}},"security":[{"bearer_auth":[]}]}},"/v1/files/{id}/publish":{"post":{"tags":["files"],"operationId":"publish_file","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UrlResponse"}}}},"404":{"description":""}},"security":[{"bearer_auth":[]}]},"delete":{"tags":["files"],"operationId":"unpublish_file","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":""},"404":{"description":""}},"security":[{"bearer_auth":[]}]}},"/v1/folders":{"get":{"tags":["folders"],"operationId":"list_folders","responses":{},"security":[{"bearer_auth":[]}]},"post":{"tags":["folders"],"operationId":"create_folder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFolderRequest"}}},"required":true},"responses":{},"security":[{"bearer_auth":[]}]}},"/v1/folders/{id}":{"get":{"tags":["folders"],"operationId":"get_folder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{},"security":[{"bearer_auth":[]}]},"delete":{"tags":["folders"],"operationId":"delete_folder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{},"security":[{"bearer_auth":[]}]},"patch":{"tags":["folders"],"operationId":"rename_folder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RenameFolderRequest"}}},"required":true},"responses":{},"security":[{"bearer_auth":[]}]}}},"components":{"schemas":{"AdminTokenItem":{"type":"object","description":"One row of `GET /v1/admin/tokens`; carries no secret material.","required":["id","name","created_at"],"properties":{"created_at":{"type":"string"},"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"revoked_at":{"type":["string","null"]}}},"AdminTokenList":{"type":"object","required":["items"],"properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/AdminTokenItem"}}}},"BillingRuleItem":{"type":"object","description":"One billing rule as served by the admin API.","required":["id","key","value","enabled","created_at"],"properties":{"created_at":{"type":"string"},"enabled":{"type":"boolean"},"id":{"type":"string","format":"uuid"},"key":{"type":"string"},"owner_id":{"type":["string","null"]},"owner_type":{"type":["string","null"]},"value":{}}},"BillingRuleList":{"type":"object","required":["items"],"properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/BillingRuleItem"}}}},"BillingStatusResponse":{"type":"object","description":"Storage usage and effective limits of the authenticated owner.","required":["used_bytes","storage_quota_bytes","max_file_bytes","rate_limit_rpm"],"properties":{"max_file_bytes":{"type":"integer","format":"int64"},"rate_limit_rpm":{"type":"integer","format":"int32","minimum":0},"storage_quota_bytes":{"type":"integer","format":"int64"},"used_bytes":{"type":"integer","format":"int64"}}},"CreateFolderRequest":{"type":"object","required":["name"],"properties":{"name":{"type":"string"}}},"CreateRuleRequest":{"type":"object","required":["key","value","enabled"],"properties":{"enabled":{"type":"boolean"},"key":{"type":"string"},"owner_id":{"type":["string","null"]},"owner_type":{"type":["string","null"]},"value":{}}},"CreateTokenRequest":{"type":"object","required":["name"],"properties":{"name":{"type":"string"}}},"CreatedAdminToken":{"type":"object","description":"Response of `POST /v1/admin/tokens`. `token` is the raw secret, returned\nexactly once; only its SHA-256 digest is ever persisted.","required":["id","name","token","created_at"],"properties":{"created_at":{"type":"string"},"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"token":{"type":"string"}}},"ErrorBody":{"type":"object","description":"JSON body shape: `{\"error\":{\"code\":...,\"message\":...}}`.","required":["error"],"properties":{"error":{"$ref":"#/components/schemas/ErrorDetail"}}},"ErrorDetail":{"type":"object","required":["code","message"],"properties":{"code":{"type":"string"},"message":{"type":"string"}}},"FileListResponse":{"type":"object","required":["items"],"properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/FileResponse"}}}},"FileResponse":{"type":"object","required":["id","name","mime_type","size_bytes","is_public","created_at"],"properties":{"created_at":{"type":"string"},"folder_id":{"type":["string","null"],"format":"uuid"},"id":{"type":"string","format":"uuid"},"is_public":{"type":"boolean"},"mime_type":{"type":"string"},"name":{"type":"string"},"size_bytes":{"type":"integer","format":"int64"}}},"FolderListResponse":{"type":"object","required":["items"],"properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/FolderResponse"}}}},"FolderResponse":{"type":"object","required":["id","name","created_at"],"properties":{"created_at":{"type":"string"},"id":{"type":"string","format":"uuid"},"name":{"type":"string"}}},"Limits":{"type":"object","description":"Effective limits for an owner after plan/billing resolution.","required":["storage_quota_bytes","max_file_bytes","rate_limit_rpm"],"properties":{"max_file_bytes":{"type":"integer","format":"int64"},"rate_limit_rpm":{"type":"integer","format":"int32","minimum":0},"storage_quota_bytes":{"type":"integer","format":"int64"}}},"Owner":{"type":"object","description":"Polymorphic owner reference (`owner_type:owner_id`).","required":["owner_type","owner_id"],"properties":{"owner_id":{"type":"string"},"owner_type":{"type":"string"}}},"Plan":{"type":"string","description":"Billing plan tiers.","enum":["Default","Premium"]},"PlanResponse":{"type":"object","required":["plan"],"properties":{"plan":{"type":"string"}}},"RenameFolderRequest":{"type":"object","required":["name"],"properties":{"name":{"type":"string"}}},"SetPlanRequest":{"type":"object","required":["plan"],"properties":{"plan":{"type":"string"}}},"UrlResponse":{"type":"object","required":["url"],"properties":{"url":{"type":"string"}}},"UsageResponse":{"type":"object","required":["used_bytes"],"properties":{"used_bytes":{"type":"integer","format":"int64"}}}},"securitySchemes":{"admin_token":{"type":"apiKey","in":"header","name":"x-admin-token","description":"Admin API token created via /v1/admin/tokens"},"bearer_auth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}